[refactor] Refactor AppUtils, SingleNote
